Restrict purchases based on inventory when attacker is self
