+++ src/components/attack/AttackForm.css
@@ -21,6 +21,18 @@
   border-radius: 4px;
   padding: 20px;
   margin: 4px;
+  position: relative;
+}
+
+.block-icon {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+
+  background: rgba(128, 128, 128, 0.826);
+  z-index: 5;
 }
 
 .icon-wrapper.active {
